+++ example-native/views/ui2.js
@@ -47,9 +47,14 @@ export default {
         label: 'ID',
         key: 'id',
         filter: false,
-        render: (val, key, row) => `<a class='button' onclick='this.dispatchEvent(new CustomEvent("testevent", { detail: ${JSON.stringify({ val, key, row })} }))'>${val}</a>`
-        // render: (val, key, row) => `<a class="button" onclick="this.dispatchEvent(new CustomEvent('testevent', { detail: { key: '${key}', val: '${val}' } }))">${val}</a>`
-        // can also fire off event - document.dispatchEvent(new CustomEvent('something', { detail: { val, row, key } }))
+        render: ({val, key, row, idx}) => {
+          const cell = val + ' is ' + (row['age'] > 5 ? '>5' : '<=5')
+          const output =
+            // do not include row as there can be too much data
+            // `<a class='button' onclick='this.dispatchEvent(new CustomEvent("testevent", { detail: ${JSON.stringify({ val, key, row, idx })} }))'>${val}</a>` // too much data if row included
+            `<a class="button" onclick="this.dispatchEvent(new CustomEvent('testevent', { detail: { key: '${key}', val: '${val}', idx: ${idx} } }))">${cell}</a>`
+          return output
+        }
       },
       {
         label: 'Name',
@@ -83,7 +88,7 @@ export default {
         age: i
       }
       for (let j=1; j<=15; j++) {
-        data['key' + j] = 'val-'+i+'-'+j
+        data['key' + j] = `r${i}-c${j}`
       }
       itemList.push(data)
     }
@@ -122,6 +127,10 @@ export default {
     }
     const testevent = (e) => {
       console.log('testevent', e.detail, e.target)
+      const { key, val, idx } = e.detail
+      console.log('ev2', table.items[idx]['age'])
+      table.items[idx]['age'] += 1
+      setItems()
     }
     
     onMounted(async () => {